AI chaser added
new class that selects closest entity and checks line of sight if there is it takes priority over entities closer but behind the wall. only one entity at the time can follow except merchants. entities use a* algorithm to find shortest path to the player and all non chasing entities move randomly. when entity steps on players tile it disappears and is treated as encounter